requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colored==1.4.3
cryptography==37.0.4
idna==3.3
loader==2017.9.11
pycparser==2.21
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.1
soupsieve==2.3.2.post1
tabulate==0.8.10
urllib3==1.26.11
